insta_spotter/requirements.txt
fastapi
uvicorn
SQLAlchemy
jinja2
python-dotenv
psycopg2-binary
asyncpg
python-jose[cryptography]
passlib[bcrypt]
requests
httpx
google-generativeai
instagrapi==2.0.0